backend/requirements.txt
aioredis==2.0.1
asgiref==3.5.2
astroid==2.12.12
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
colorama==0.4.6
constantly==15.1.0
coverage==6.5.0
coveralls==3.3.1
cryptography==38.0.1
daphne==4.0.0
Deprecated==1.2.13
dill==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-seed==0.3.1
docopt==0.6.2
Faker==15.1.3
hiredis==2.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
msgpack==1.0.4
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==21.3
platformdirs==2.5.2
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pylint==2.15.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
toposort==1.7
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
zope.interface==5.5.0
